backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
pgvector==0.2.5
alembic==1.13.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.7.4
httpx==0.27.0
celery==5.3.6
redis==5.0.4
numpy==1.26.4
prometheus-client==0.20.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-jaeger==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-requests==0.46b0
sentence-transformers==2.7.0
requests==2.32.3
